added no authorization messages for non-admin or non-dentist staff

+++ PatientDetails.cs
@@ -36,9 +36,29 @@ namespace DentalAppointmentandInformationSystem
 
         private void editPatientBtn_Click(object sender, EventArgs e)
         {
-            EditPatient editptnt = new EditPatient();
-            editptnt.Show();
-            this.Hide();
+            constring.Open();
+            SqlCommand cmd = new SqlCommand("SELECT * FROM Staff WHERE employee_num =" + int.Parse(v.getsetloggedIn), constring);
+            SqlDataReader reader1;
+            reader1 = cmd.ExecuteReader();
+            if (reader1.Read())
+            {
+                if (reader1.GetValue(7).ToString().Equals("Dentist") || reader1.GetValue(7).ToString().Equals("Administrator"))
+                {
+                    constring.Close();
+                    EditPatient editptnt = new EditPatient();
+                    editptnt.Show();
+                    this.Hide();
+                }
+                else
+                {
+                    constring.Close();
+                    MessageBox.Show("You do not have the authorization to edit patient records!");
+                }
+            }
+            else
+            {
+                MessageBox.Show("NO DATA FOUND");
+            }
         }
 
         private void displayPatientDetails()
@@ -209,38 +229,57 @@ namespace DentalAppointmentandInformationSystem
 
         private void deletePtnt_Click(object sender, EventArgs e)
         {
-            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this patient? \nAll records of the patient will be deleted!", "Confirm Delete", MessageBoxButtons.YesNo);
-            if (dialogResult == DialogResult.Yes)
+            constring.Open();
+            SqlCommand cmd = new SqlCommand("SELECT * FROM Staff WHERE employee_num =" + int.Parse(v.getsetloggedIn), constring);
+            SqlDataReader reader1;
+            reader1 = cmd.ExecuteReader();
+            if (reader1.Read())
             {
-                constring.Open();
-                string query = "DELETE FROM Appointment WHERE patient_id =" + v.getsetpatientSelected;
-                SqlCommand cmd = new SqlCommand(query, constring);
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-                string query2 = "DELETE FROM Record WHERE patient_id =" + v.getsetpatientSelected;
-                SqlCommand cmd2 = new SqlCommand(query2, constring);
-                cmd2.CommandText = query2;
-                cmd2.ExecuteNonQuery();
-                string query3 = "DELETE FROM History WHERE patient_id =" + v.getsetpatientSelected;
-                SqlCommand cmd3 = new SqlCommand(query3, constring);
-                cmd3.CommandText = query3;
-                cmd3.ExecuteNonQuery();
-                string query4 = "DELETE FROM Patient WHERE patient_id =" + v.getsetpatientSelected;
-                SqlCommand cmd4 = new SqlCommand(query4, constring);
-                cmd4.CommandText = query4;
-                if (cmd4.ExecuteNonQuery() == 1)
+                if (reader1.GetValue(7).ToString().Equals("Dentist") || reader1.GetValue(7).ToString().Equals("Administrator"))
                 {
-                    MessageBox.Show("Patient Deleted Successfully!");
-                    constring.Close();
-                    Patients ptnt = new Patients();
-                    ptnt.Show();
-                    this.Hide();
+                    DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this patient? \nAll records of the patient will be deleted!", "Confirm Delete", MessageBoxButtons.YesNo);
+                    if (dialogResult == DialogResult.Yes)
+                    {
+                        string query = "DELETE FROM Appointment WHERE patient_id =" + v.getsetpatientSelected;
+                        SqlCommand cmdd = new SqlCommand(query, constring);
+                        cmdd.CommandText = query;
+                        cmdd.ExecuteNonQuery();
+                        string query2 = "DELETE FROM Record WHERE patient_id =" + v.getsetpatientSelected;
+                        SqlCommand cmd2 = new SqlCommand(query2, constring);
+                        cmd2.CommandText = query2;
+                        cmd2.ExecuteNonQuery();
+                        string query3 = "DELETE FROM History WHERE patient_id =" + v.getsetpatientSelected;
+                        SqlCommand cmd3 = new SqlCommand(query3, constring);
+                        cmd3.CommandText = query3;
+                        cmd3.ExecuteNonQuery();
+                        string query4 = "DELETE FROM Patient WHERE patient_id =" + v.getsetpatientSelected;
+                        SqlCommand cmd4 = new SqlCommand(query4, constring);
+                        cmd4.CommandText = query4;
+                        if (cmd4.ExecuteNonQuery() == 1)
+                        {
+                            MessageBox.Show("Patient Deleted Successfully!");
+                            constring.Close();
+                            Patients ptnt = new Patients();
+                            ptnt.Show();
+                            this.Hide();
+                        }
+                        else
+                        {
+                            MessageBox.Show("DATA NOT DELETED SUCCESSFULLY");
+                        }
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("DATA NOT DELETED SUCCESSFULLY");
+                    MessageBox.Show("You do not have the authorization to delete patient records!");
+                    constring.Close();
                 }
             }
+            else
+            {
+                MessageBox.Show("NO DATA FOUND");
+            }
+            
         }
     }
 }
