+++ ServicesList.cs
@@ -37,36 +37,76 @@ namespace DentalAppointmentandInformationSystem
 
         private void editBtn_Click(object sender, EventArgs e)
         {
-            v.getsetserviceSelected = serviceID.Text;
-            Services mainForm;
-            mainForm = (Services)this.FindForm();
-            mainForm.editVisible();
+            constring.Open();
+            SqlCommand cmd = new SqlCommand("SELECT * FROM Staff WHERE employee_num =" + int.Parse(v.getsetloggedIn), constring);
+            SqlDataReader reader1;
+            reader1 = cmd.ExecuteReader();
+            if (reader1.Read())
+            {
+                if (reader1.GetValue(7).ToString().Equals("Dentist") || reader1.GetValue(7).ToString().Equals("Administrator"))
+                {
+                    constring.Close();
+                    v.getsetserviceSelected = serviceID.Text;
+                    Services mainForm;
+                    mainForm = (Services)this.FindForm();
+                    mainForm.editVisible();
+                }
+                else
+                {
+                    MessageBox.Show("You do not have the authorization to edit services!");
+                    constring.Close();
+                }
+            }
+            else
+            {
+                MessageBox.Show("NO DATA FOUND");
+            }
+            
         }
 
         private void deleteBtn_Click(object sender, EventArgs e)
         {
-            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this service?", "Confirm Delete", MessageBoxButtons.YesNo);
-            if (dialogResult == DialogResult.Yes)
+            constring.Open();
+            SqlCommand cmd = new SqlCommand("SELECT * FROM Staff WHERE employee_num =" + int.Parse(v.getsetloggedIn), constring);
+            SqlDataReader reader1;
+            reader1 = cmd.ExecuteReader();
+            if (reader1.Read())
             {
-                v.getsetserviceSelected = serviceID.Text;
-                constring.Open();
-                string query = "DELETE FROM Service WHERE service_id ='" + v.getsetserviceSelected + "';";
-                SqlCommand cmd = new SqlCommand(query, constring);
-                cmd.CommandText = query;
-                if (cmd.ExecuteNonQuery() == 1)
+                if (reader1.GetValue(7).ToString().Equals("Dentist") || reader1.GetValue(7).ToString().Equals("Administrator"))
                 {
-                    MessageBox.Show("Service Deleted Successfully!");
-                    constring.Close();
-                    Services srvcs = new Services();
-                    srvcs.Show();
-                    this.ParentForm.Hide();
+                    DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this service?", "Confirm Delete", MessageBoxButtons.YesNo);
+                    if (dialogResult == DialogResult.Yes)
+                    {
+                        v.getsetserviceSelected = serviceID.Text;
+                        string query = "DELETE FROM Service WHERE service_id ='" + v.getsetserviceSelected + "';";
+                        SqlCommand cmdd = new SqlCommand(query, constring);
+                        cmdd.CommandText = query;
+                        if (cmdd.ExecuteNonQuery() == 1)
+                        {
+                            MessageBox.Show("Service Deleted Successfully!");
+                            constring.Close();
+                            Services srvcs = new Services();
+                            srvcs.Show();
+                            this.ParentForm.Hide();
+                        }
+                        else
+                        {
+                            MessageBox.Show("Service cannot be deleted because it is included in a patient's record!");
+                        }
+
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("Service cannot be deleted because it is included in a patient's record!");
+                    MessageBox.Show("You do not have the authorization to delete services!");
+                    constring.Close();
                 }
-                
             }
+            else
+            {
+                MessageBox.Show("NO DATA FOUND");
+            }
+            
         }
     }
 }
